Moved Player and Missile classes to separate files
